Try to fix lint
